+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway/events.ex
@@ -8,10 +8,9 @@ defmodule RigInboundGateway.Events do
   alias Rig.Subscription
   alias RigCloudEvents.CloudEvent
 
-  @spec welcome_event() :: CloudEvent.t()
-  def welcome_event do
-    connection_pid = self()
-    connection_token = Connection.Codec.serialize(connection_pid)
+  @spec welcome_event(pid | nil) :: CloudEvent.t()
+  def welcome_event(pid \\ self()) do
+    connection_token = Connection.Codec.serialize(pid)
 
     rig_event(
       "rig.connection.create",