+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway_web/router.ex
@@ -19,6 +19,14 @@ defmodule RigInboundGatewayWeb.Router do
         # up using the Cowboy dispatch configuration; see the `config.exs` file.
       end
 
+      # Unlike SSE & WS handlers, the LP handler is implemented using plug
+      scope "/connection/longpolling" do
+        subscription_url = "/:connection_id/subscriptions"
+        get("/", LongpollingController, :handle_connection)
+        options(subscription_url, SubscriptionController, :handle_preflight)
+        put(subscription_url, SubscriptionController, :set_subscriptions)
+      end
+
       scope "/connection/ws" do
         subscription_url = "/:connection_id/subscriptions"
         options(subscription_url, SubscriptionController, :handle_preflight)